+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Zscaler Internet Access (ZIA) Ansible Collection Changelog
 
+## v2.0.2 (May, 26 2025)
+
+### Notes
+
+- Python Versions: **v3.9, v3.10, v3.11**
+
+### Bug Fixes
+
+[#75](https://github.com/zscaler/ziacloud-ansible/pull/75) - Added new `source_ip_groups` attribute to `zia_url_filtering_rules`
+[#75](https://github.com/zscaler/ziacloud-ansible/pull/75) - Added new `nw_applications` attribute to `zia_cloud_firewall_rule` and `zia_forwarding_control_rule`.
+[#75](https://github.com/zscaler/ziacloud-ansible/pull/75) - Fixed drift with the attribute `ip_addresses` within the resource `zia_cloud_firewall_ip_source_groups`
+
 ## v2.0.1 (May, 12 2025)
 
 ### Notes
@@ -8,7 +20,7 @@
 
 ### Bug Fixes
 
-[#72](https://github.com/zscaler/ziacloud-ansible/issues/72) - Fixed `all_collect_items` pagination argument count
+[#72](https://github.com/zscaler/ziacloud-ansible/pull/72) - Fixed `all_collect_items` pagination argument count
 
 ## 2.0.0 (May, 6 2025) - BREAKING CHANGES